Dragen: collecting sample names and reporting the number of found samples once
